+++ backend/users/Touir.ExpensesManager.Users/Infrastructure/Options/RabbitMQOptions.cs
@@ -1,0 +1,10 @@
+namespace Touir.ExpensesManager.Users.Infrastructure.Options
+{
+    public class RabbitMQOptions
+    {
+        public string? HostName { get; set; }
+        public int Port { get; set; }
+        public string? UserName { get; set; }
+        public string? Password { get; set; }
+    }
+}