+++ backend/users/Touir.ExpensesManager.Users/Services/Contracts/IRabbitMQService.cs
@@ -1,0 +1,9 @@
+using RabbitMQ.Client;
+
+namespace Touir.ExpensesManager.Users.Services.Contracts
+{
+    public interface IRabbitMQService
+    {
+        IConnection GetConnection();
+    }
+}